requirements.txt
fastmcp
mcp[cli]
azure-ai-inference
azure-core
fastapi
uv
uvicorn
openai
python-dotenv
flake8
isort